+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Following Semantic Versioning 2.
 
 ## Version 0.0.10 (MINOR)
 - In show Challenge, hide "Keywords" and "Proposed solutions" when they are empty.
+- In show Problem, hide "Keywords" and "Proposed solutions" when they are empty.
 - In show Solution, hide "Keywords", "Beneficiaries", "Requeriments" and "Indicators" when they are empty.
 - Upgrade module's Ruby to 2.7.5
 - Fix: hide Challenges, Problems and Solutions from global search when its Participatory space is not visible